ai assistant/backend/requirements_dev.txt
# Development requirements (uses SQLite instead of PostgreSQL)
# Use this if psycopg2-binary installation fails on Windows

Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
# psycopg2-binary==2.9.9  # Commented out - use SQLite for dev
python-dotenv==1.0.0
openai==1.3.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0

